W-12207110: Wrap an IllegalArgumentException in a MessagingException at ForeachRouter in order to fix an event stuck (#12066)
(cherry picked from commit 2511088f1170b4ab03b29bb13b52dda8af8be0a4)
